% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_generation.R
\name{natlogo_generation}
\alias{natlogo_generation}
\title{Generate Logo for the natverse package}
\usage{
natlogo_generation()
}
\description{
This will generate logo for natverse package based on the png image created by Alex Bates
Further produce a hex sticker based on the package \code{hexSticker}
Adapted and modified from the following sources : \url{https://github.com/Bioconductor/BiocStickers}
}
\examples{
\dontrun{
library(hexSticker)
library(ggplot2)
library(here)
natlogo_generation()
}

}
