% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natverse-package.R
\docType{package}
\name{natverse-package}
\alias{natverse}
\alias{natverse-package}
\title{natverse: Install all NeuroAnatomy Toolbox packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functionality for installing and updating all nat packages and friends in one go. This includes both CRAN and github only packages. Other functionalities like checking conflicts are also provided. Inspired by tidyverse and dtupdate packages.
}
\details{
The key function for end users is \code{\link{natverse_update}} which will
allow you to conveniently update all natverse packages and their
dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/natverse}
  \item \url{https://natverse.github.io/nat/}
  \item Report bugs at \url{https://github.com/natverse/natverse/issues}
}

}
\author{
\strong{Maintainer}: Sridhar Jagannathan \email{j.sridharrajan@gmail.com} (\href{https://orcid.org/0000-0002-2078-1145}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

Other contributors:
\itemize{
  \item Alexander Bates \email{abates@mrc-lmb.cam.ac.uk} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID}) [contributor]
}

}
\keyword{internal}
