% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{natverse_conflicts}
\alias{natverse_conflicts}
\title{Detects the conflicts between natverse and other packages}
\usage{
natverse_conflicts()
}
\description{
This function lists all the conflicts between packages in the natverse
and other packages have been already loaded n the environment.
Adapted and modified from conflicts function in 'tidyverse' package
}
\examples{
natverse_conflicts()
}
