% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{natverse_deps}
\alias{natverse_deps}
\title{print status of natverse dependencies
This will check to see if all natverse packages (and optionally (if
\code{dependencies = TRUE}), their dependencies )}
\usage{
natverse_deps(dependencies = TRUE)
}
\arguments{
\item{dependencies}{If \code{TRUE}, will also check all dependencies of natverse packages.}
}
\value{
A character vector containing packages that are either missing or are not up-to-date.
}
\description{
print status of natverse dependencies
This will check to see if all natverse packages (and optionally (if
\code{dependencies = TRUE}), their dependencies )
}
\examples{
\dontrun{
natverse_deps()
}
}
