% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{natverse_update}
\alias{natverse_update}
\title{Check status of natverse packages and (optionally) update if necessary}
\usage{
natverse_update(update = FALSE, recursive = TRUE, dependencies = NA,
  ...)
}
\arguments{
\item{update}{If \code{TRUE}, will actually update the packages}

\item{recursive}{If \code{TRUE}, will also check all dependencies of natverse packages.}

\item{dependencies}{this parameter is passed onto function \code{\link[remotes]{update_packages}},
indicating which dependencies
to update here NA stands for "Depends", "Imports" and "LinkingTo" .}

\item{...}{extra arguments to pass to \code{\link[remotes]{update_packages}}.}
}
\description{
This will check to see if all natverse packages (and optionally (if
\code{recursive = TRUE}), their dependencies ) are up-to-date, and will
update them (optionally (if \code{update = TRUE})) if they are missing or are not up-to-date!
}
\examples{
natverse_update()
}
