% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natverse-package.R
\docType{package}
\name{natverse-package}
\alias{natverse}
\alias{natverse-package}
\title{natverse: Install all NeuroAnatomy Toolbox packages}
\description{
Provides functionality for installing and updating
    all nat packages and friends in one go. This includes both CRAN and
    github only packages. Other functionalities like checking conflicts
    are also provided. Inspired from tidyverse and dtupdate packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/natverse}
  \item Report bugs at \url{https://github.com/natverse/natverse/issues}
}

}
\author{
\strong{Maintainer}: Sridhar Jagannathan \email{j.sridharrajan@gmail.com} (0000-0002-2078-1145)

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)
}

Other contributors:
\itemize{
  \item Alex Bates \email{abates@mrc-lmb.cam.ac.uk} (0000-0002-1195-0445) [contributor]
}

}
\keyword{internal}
