% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{natverse_update}
\alias{natverse_update}
\title{Check status of natverse packages and (optionally) update if necessary}
\usage{
natverse_update(recursive = FALSE, source = c("CRAN", "GITHUB"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of natverse
packages.}

\item{source}{set the source of updates to 'CRAN' or 'GITHUB' natverse
packages. The default checks both.}
}
\value{
A \code{tibble} detailing dependencies and their status.
}
\description{
This will check to see if all natverse packages (and optionally (if
\code{recursive = TRUE}), their dependencies ) are up-to-date, and will
provide the command to perform the installation in one go! Adapted and
modified from the following sources : \code{update} function in 'tidyverse'
package, \code{github_update} function in 'dtupdate' package
}
\examples{
natverse_update()
}
